Introduce new feature for rapidly capturing points
When the left mouse button is pressed, points are captured as the
mouse moves. Rapid capturing stop when the button is released. The +
and - buttons on the main toolbar set the minimum distance between two
points during rapid capture. That is, a new point is generated only if
it is a certain distance away from the previous point. The rapid
capture level (set by + and - buttons) goes from 1 to 10. These
correspond to the minimum distances between the points from 1.0 to
10.0 pixels, respectively.